update bmds 2 vs 3
