+++ frontend/bmd/bmds2/components/BaseModal.js
