+++ frontend/bmd/bmds2/components/EditableModalFooter.js
