+++ frontend/bmd/bmds2/components/OutputFigure.js
