+++ frontend/bmd/bmds2/components/OutputModal.js
