+++ frontend/bmd/bmds2/components/Recommendation.js
@@ -2,7 +2,7 @@ import _ from "lodash";
 import React from "react";
 import PropTypes from "prop-types";
 
-import {asLabel} from "bmd/models/bmr";
+import {asLabel} from "../models/bmr";
 
 import RecommendationNotes from "./RecommendationNotes";
 import RecommendationTable from "./RecommendationTable";
