+++ frontend/bmd/bmds2/components/RecommendationNotes.js
