+++ frontend/bmd/bmds2/constants.js
