+++ frontend/bmd/bmds3/index.js
@@ -1,0 +1,6 @@
+import React from "react";
+import ReactDOM from "react-dom";
+
+export default function(el, config) {
+    ReactDOM.render(<h1>Hello bmds3.</h1>, el);
+}
