+++ hawc/apps/bmd/templates/bmd/session_detail.html
@@ -47,7 +47,7 @@
 $(document).ready(function () {
   const el = document.getElementById('bmd'),
     config = JSON.parse(document.getElementById('config').text);
-  window.app.bmdStartup(el, config);
+  window.app.bmds2Startup(el, config);
 });
 </script>
 {% endblock %}
