+++ hawc/apps/bmd/templates/bmd/session_form.html
@@ -45,7 +45,7 @@
 $(document).ready(function(){
   const el = document.getElementById('bmd'),
     config = JSON.parse(document.getElementById('config').text);
-  window.app.bmdStartup(el, config);
+  window.app.bmds2Startup(el, config);
 });
 </script>
 {% endblock %}
